+++ mobile-eyuel/task_6/lib/features/ecommerce/data/datasource/product_remote_datasource.dart
@@ -2,8 +2,10 @@ import 'dart:convert';
 
 import 'package:http/http.dart' as http;
 import 'package:http_parser/http_parser.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 import '../../../../core/errors/exceptions.dart';
+import '../../../../core/errors/failure.dart';
 import '../../domain/entities/product.dart';
 import '../models/product_model.dart';
 
@@ -17,21 +19,29 @@ abstract class ProductRemoteDatasource {
 
 class ProductRemoteDataSourceImpl implements ProductRemoteDatasource {
   final http.Client client;
+  final SharedPreferences sharedPreferences;
 
   static const String baseUrl =
-      'https://g5-flutter-learning-path-be.onrender.com/api/v1/products';
+      'https://g5-flutter-learning-path-be.onrender.com/api/v3/products';
 
-  ProductRemoteDataSourceImpl({required this.client});
+  static const String tTokenKey = 'access_token';
+
+  ProductRemoteDataSourceImpl(
+      {required this.client, required this.sharedPreferences});
 
   @override
   Future<void> createProduct(ProductEntity product) async {
+    final token = sharedPreferences.getString(tTokenKey);
+    if (token == null) {
+      throw const AuthFailure();
+    }
     var request = http.MultipartRequest('POST', Uri.parse(baseUrl));
+    request.headers
+        .addAll({'Authorization': token, 'Content-Type': 'application/json'});
     request.fields.addAll({
-      // 'id': product.id,
       'name': product.name,
       'description': product.description,
-      'price': product.price.toString(),
-      // 'imageUrl': product.imageUrl,
+      'price': product.price.toString()
     });
     request.files.add(await http.MultipartFile.fromPath(
         'image', product.imageUrl,
@@ -48,8 +58,12 @@ class ProductRemoteDataSourceImpl implements ProductRemoteDatasource {
 
   @override
   Future<void> deleteProduct(String id) async {
+    final token = sharedPreferences.getString(tTokenKey);
+    if (token == null) {
+      throw const AuthFailure();
+    }
     final response = await client.delete(Uri.parse('$baseUrl/$id'),
-        headers: {'Content-Type': 'application/json'});
+        headers: {'Authorization': token, 'Content-Type': 'application/json'});
 
     if (response.statusCode == 200) {
       return;
@@ -60,9 +74,13 @@ class ProductRemoteDataSourceImpl implements ProductRemoteDatasource {
 
   @override
   Future<List<ProductModel>> getAllProducts() async {
+    final token = sharedPreferences.getString(tTokenKey);
+    if (token == null) {
+      throw const AuthFailure();
+    }
     final response = await client.get(
       Uri.parse(baseUrl),
-      headers: {'Content-Type': 'application/json'},
+      headers: {'Authorization': token, 'Content-Type': 'application/json'},
     );
 
     if (response.statusCode == 200) {
@@ -77,9 +95,13 @@ class ProductRemoteDataSourceImpl implements ProductRemoteDatasource {
 
   @override
   Future<ProductModel> getSingleProduct(String id) async {
+    final token = sharedPreferences.getString(tTokenKey);
+    if (token == null) {
+      throw const AuthFailure();
+    }
     final response = await client.get(
       Uri.parse('$baseUrl/$id'),
-      headers: {'Content-Type': 'application/json'},
+      headers: {'Authorization': token, 'Content-Type': 'application/json'},
     );
 
     if (response.statusCode == 200) {
@@ -91,8 +113,12 @@ class ProductRemoteDataSourceImpl implements ProductRemoteDatasource {
 
   @override
   Future<void> updateProduct(ProductModel product) async {
+    final token = sharedPreferences.getString(tTokenKey);
+    if (token == null) {
+      throw const AuthFailure();
+    }
     final response = await client.put(Uri.parse('$baseUrl/${product.id}'),
-        headers: {'Content-Type': 'application/json'},
+        headers: {'Authorization': token, 'Content-Type': 'application/json'},
         body: jsonEncode(product.tojson()));
 
     if (response.statusCode == 200) {
